+++ testsuite/mypyc-dir/partial-package/package/sub.py
@@ -1,0 +1,3 @@
+import os
+
+SUB_VALUE = 42