+++ testsuite/mypyc-dir/plain-init-package/package/__init__.py
@@ -1,0 +1,3 @@
+import pathlib
+from . import sub
+from . import mod